add logic to send emails, and finish launching workspace
