+++ webapp/channels/src/components/preparing_workspace/plugins.tsx
@@ -10,7 +10,6 @@ import MultiSelectCards from 'components/common/multi_select_cards';
 
 import GithubSVG from 'components/common/svg_images_components/github_svg';
 import GitlabSVG from 'components/common/svg_images_components/gitlab_svg';
-import CelebrateSVG from 'components/common/svg_images_components/celebrate_svg';
 import JiraSVG from 'components/common/svg_images_components/jira_svg';
 import ZoomSVG from 'components/common/svg_images_components/zoom_svg';
 import TodoSVG from 'components/common/svg_images_components/todo_svg';
@@ -46,32 +45,19 @@ const Plugins = (props: Props) => {
         className += ' ' + props.className;
     }
 
-    let title = (
+    const title = (
         <FormattedMessage
-            id={'onboarding_wizard.cloud_plugins.title'}
-            defaultMessage='Welcome to Mattermost!'
+            id={'onboarding_wizard.self_hosted_plugins.title'}
+            defaultMessage='What tools do you use?'
         />
     );
-    let description = (
+
+    const description = (
         <FormattedMessage
-            id={'onboarding_wizard.cloud_plugins.description'}
-            defaultMessage={'Mattermost is better when integrated with the tools your team uses for collaboration. Popular tools are below, select the ones your team uses and we\'ll add them to your workspace. Additional set up may be needed later.'}
+            id={'onboarding_wizard.self_hosted_plugins.description'}
+            defaultMessage={'Choose the tools you work with, and we\'ll add them to your workspace. Additional set up may be needed later.'}
         />
     );
-    if (props.isSelfHosted) {
-        title = (
-            <FormattedMessage
-                id={'onboarding_wizard.self_hosted_plugins.title'}
-                defaultMessage='What tools do you use?'
-            />
-        );
-        description = (
-            <FormattedMessage
-                id={'onboarding_wizard.self_hosted_plugins.description'}
-                defaultMessage={'Choose the tools you work with, and we\'ll add them to your workspace. Additional set up may be needed later.'}
-            />
-        );
-    }
 
     return (
         <CSSTransition
@@ -94,16 +80,6 @@ const Plugins = (props: Props) => {
                     {props.previous}
                     <Title>
                         {title}
-                        {!props.isSelfHosted && (
-                            <div className='subtitle'>
-                                <CelebrateSVG/>
-                                <FormattedMessage
-                                    id={'onboarding_wizard.cloud_plugins.subtitle'}
-                                    defaultMessage='(almost there!)'
-                                />
-                            </div>
-
-                        )}
                     </Title>
                     <Description>{description}</Description>
                     <PageBody>
