+++ webapp/channels/src/components/preparing_workspace/preparing_workspace.scss
@@ -133,6 +133,7 @@
 
     // centering
     margin: 0 auto;
+    margin-left: 300px;
 }
 
 @media screen and (max-width: 768px) {
