+++ webapp/channels/src/sass/base/_structure.scss
@@ -19,7 +19,6 @@ body {
 
     &.admin-onboarding {
         background-image: url('images/admin-onboarding-background.jpg');
-        background-position: 50%;
         background-repeat: no-repeat;
         background-size: cover;
     }
